Did html/css for playlists and home. app.py slightly changed
HTML/CSS for playlists(depending on scheme and color can still be work in progress)
css for liked songs still needs to be worked on
app.py changed to accomidate the home page and welcome. But now playlists and liked_songs do not work right. So that still needs to be worked on.